+++ AlarmServer/ValueChangeEvent.cs
@@ -1,24 +1,20 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AlarmServer
 {
     public class ValueChangeEvent : EventModel
     {
-        public int Value { get; }
+        public long Value { get; }
         public string Name { get; }
 
-        public ValueChangeEvent(DateTime timestamp, string name, int value) : 
+        public ValueChangeEvent(DateTime timestamp, string name, long value) : 
             base(timestamp, name + " " + ToString(value))
         {
             Name = name;
             Value = value;
         }
 
-        static string ToString(int value)
+        static string ToString(long value)
         {
             return value.ToString();
         }
